cpufreq: cpu-boost: don't boost migrations with load below the threshold
Even if we don't use load_based_syncs (there seems to be a bug that make
min_freq stuck at random frequencies) we can prevent this feature
to boost the target cpu if the task load is below the migration
threshold. We might save some mA down the road with this.

Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>